Migrate to integrations.deepspeed rather than modeling_utils
